Update App.tsx
Making the graph continuously update instead of having to click it a
bunch of times. Also the kind of graph we want to serve as visual here is
kind of a continuously updating line graph whose y axis is the stock’s
top_ask_price and the x-axis is the timestamp of the stock.